backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.0
django-filter==23.1
python-dotenv==1.0.1
psycopg2-binary==2.9.3